move logic out of controller
